Update details shown for movies and save sort pref
Add more data to the movies downloaded from JSON. Displaying more of
this data in the DetailFragment. Also, added code for scalling the image
on the DetailFragment so it properly fills the ImageView while keeping
scalled properly.

Added sort by filter to the movie listing. The filter is saved as a
preference. The preference can be set from the SettingsActivity or by
choosing a menu option from the MenuBar. Upon selecting a different sort
option the option is saved in the settings and the menu options change
accordingly.